requirements.txt
# Slack & Discord
slack-bolt>=1.23.0
discord.py>=2.4.0

# HTTP & Scheduling
requests>=2.31.0
APScheduler>=3.10.0
pymongo>=4.4.0
python-dotenv>=1.0.0